Add Target equality and integration for spell casting
NEED HEAVY CLEANSING AND REWORKING SINCE IT MESSY RN

Implemented __eq__ and __hash__ methods for the Target class to allow instances to be used in hashable data structures. Integrated the Target class into the spell-casting mechanism by using a set for unique target tracking. This enhances the efficiency and flexibility of spell management, allowing better interaction with spell-casting operations.